+++ src/pages/design-sprint.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Layout from '../components/modules/layout'
 import styled from '../util/styled-components'
 import SplitContent from '../components/common/splitcontent'
-import HeroHeader from '../components/common/heroheader'
+import HeroHeader from '../components/common/hero-header'
 import svgPostIts from '../images/design-sprint-post-its.svg'
 import svgExplanation from '../images/design-sprint-explanation.svg'
 import svgMonTues from '../images/design-sprint-monday-tuesday.svg'
